\name{rplotlib-package}
\alias{rplotlib-package}
\alias{rplotlib}
\docType{package}
\title{
Plot Functions
}
\description{
Builds on base plotting functions.
}
\details{
\tabular{ll}{
Package: \tab rplotlib\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-20\cr
License: \tab Public Domain\cr
}

}
\author{
Ben Miller

Maintainer: Ben Miller <bsmiller25@gmail.com>
}
\keyword{ package }
